requirements.txt
torch>=1.4
torchvision>=0.5.0
opencv-python>=4.0
scipy>=1.4.0
h5py>=2.10
pillow>=7.0.0
imageio>=1.18
nni>=2.0
numpy
yacs
easydict
